changed float to double for percentages

+++ src/main/java/com/napier/sem/ReportUtil.java
@@ -704,9 +704,9 @@ public class ReportUtil {
         return result;
     }
 
-    private static float roundedPercentage(float value, long total){
-        float result = ((float)value) / total * 100;
-        result = ((float)Math.round(result * 100)) / 100;
+    private static float roundedPercentage(double value, long total){
+        double result = ((double)value) / total * 100;
+        result = ((double)Math.round(result * 100)) / 100;
         return result;
     }
